+++ packages/components/pie-list/test/visual/pie-list.spec.ts
@@ -23,6 +23,11 @@ const visualStories: { storyId: string, snapshotName: string }[] = [
     { storyId: 'list--borders', snapshotName: 'PieList - Borders between items' },
     { storyId: 'list--switch-selection', snapshotName: 'PieList - Switch selection list' },
     { storyId: 'list--link-list', snapshotName: 'PieList - Link list' },
+    { storyId: 'list--button-list', snapshotName: 'PieList - Button list' },
+    { storyId: 'list--button-list-disabled', snapshotName: 'PieList - Button list (disabled)' },
+    { storyId: 'list--item-height-compact', snapshotName: 'PieList - Item height (compact, primary text only)' },
+    { storyId: 'list--item-height-primary-and-secondary', snapshotName: 'PieList - Item height (primary and secondary text)' },
+    { storyId: 'list--item-height-primary-only', snapshotName: 'PieList - Item height (primary text only)' },
 ];
 
 test.describe('PieList - Visual tests', () => {